// packages/boundingbox/src/types.d.ts
export interface BaseAnnotation {
    id: string;
    label: string;
    color?: string;
    attributes?: Record<string, any>;
}
/**
 * Bounding box annotation
 */
export interface BoundingBox extends BaseAnnotation {
    x: number;
    y: number;
    width: number;
    height: number;
}
/**
 * Polygon annotation for segmentation masks
 */
export interface PolygonAnnotation extends BaseAnnotation {
    points: {
        x: number;
        y: number;
    }[];
}
/**
 * Union type for all annotation types
 */
export type Annotation = BoundingBox | PolygonAnnotation;
/**
 * Image information required for coordinate transformations
 */
export interface ImageInfo {
    width: number;
    height: number;
    src?: string;
    alt?: string;
}